Pairs: * Move Package Structs from Numbers Module to Pairs Module
