+++ pairs/src/main/java/mathtools/pairs/BytePairFixed.java
@@ -1,4 +1,4 @@
-package mathtools.numbers.structs;
+package mathtools.pairs;
 
 /** A pair of primitive 8 bit integers that are immutable.
  * @author DK96-OS : 2023
