+++ pairs/src/main/java/mathtools/pairs/LongPairFixed.java
@@ -1,4 +1,4 @@
-package mathtools.numbers.structs;
+package mathtools.pairs;
 
 /** A pair of primitive 64-bit integers that are immutable.
  * @author DK96-OS : 2023
