+++ pairs/src/main/java/mathtools/pairs/package-info.java
@@ -2,4 +2,4 @@
  * Numerical Data Structures
  * @author DK96-OS
  */
-package mathtools.numbers.structs;
+package mathtools.pairs;
